:muscle: Move `core/action` into `action/core`
