+++ denops/gin/action/core/registry.ts
