+++ denops/gin/main.ts
@@ -1,6 +1,6 @@
 import type { Denops } from "https://deno.land/x/denops_std@v4.0.0/mod.ts";
 
-import { main as mainAction } from "./core/action/main.ts";
+import { main as mainAction } from "./action/core/main.ts";
 import { main as mainBare } from "./command/bare/main.ts";
 import { main as mainBuffer } from "./command/buffer/main.ts";
 import { main as mainProxy } from "./proxy/main.ts";
